+++ core/src/main/scala/de/sciss/fscape/stream/impl/WindowedLogicImpl.scala
@@ -30,7 +30,7 @@ trait WindowedLogicImpl[S <: Shape] extends ChunkImpl[S] {
     * @return the number of frames to write to the internal window buffer
     *         (becomes `writeToWinRemain`)
     */
-  protected def startNextWindow(inOff: Int): Int
+  protected def startNextWindow(inOff: Int): Long
 
   /** If crucial inputs have been closed. */
   protected def inputsEnded: Boolean
@@ -41,7 +41,7 @@ trait WindowedLogicImpl[S <: Shape] extends ChunkImpl[S] {
     * @param writeToWinOff  current offset into internal window buffer
     * @param chunk          number of frames to copy
     */
-  protected def copyInputToWindow(inOff: Int, writeToWinOff: Int, chunk: Int): Unit
+  protected def copyInputToWindow(inOff: Int, writeToWinOff: Long, chunk: Int): Unit
 
   /** Called when the internal window buffer is full, in order to
     * proceed to the next phase of copying from window to output.
@@ -53,16 +53,16 @@ trait WindowedLogicImpl[S <: Shape] extends ChunkImpl[S] {
     * @return the number of frames available for sending through `copyWindowToOutput`
     *         (this becomes `readFromWinRemain`).
     */
-  protected def processWindow(writeToWinOff: Int /* , flush: Boolean */): Int
+  protected def processWindow(writeToWinOff: Long): Long
 
-  protected def copyWindowToOutput(readFromWinOff: Int, outOff: Int, chunk: Int): Unit
+  protected def copyWindowToOutput(readFromWinOff: Long, outOff: Int, chunk: Int): Unit
 
   // ---- impl ----
 
-  private[this] final var writeToWinOff     = 0
-  private[this] final var writeToWinRemain  = 0
-  private[this] final var readFromWinOff    = 0
-  private[this] final var readFromWinRemain = 0
+  private[this] final var writeToWinOff     = 0L
+  private[this] final var writeToWinRemain  = 0L
+  private[this] final var readFromWinOff    = 0L
+  private[this] final var readFromWinRemain = 0L
   private[this] final var isNextWindow      = true
 
   @inline
@@ -80,7 +80,7 @@ trait WindowedLogicImpl[S <: Shape] extends ChunkImpl[S] {
         // logStream(s"startNextWindow(); writeToWinRemain = $writeToWinRemain")
       }
 
-      val chunk     = math.min(writeToWinRemain, inRemain)
+      val chunk     = math.min(writeToWinRemain, inRemain).toInt
       val flushIn   = flushIn0 && writeToWinOff > 0
       if (chunk > 0 || flushIn) {
         // logStream(s"writeToWindow(); inOff = $inOff, writeToWinOff = $writeToWinOff, chunk = $chunk")
@@ -105,7 +105,7 @@ trait WindowedLogicImpl[S <: Shape] extends ChunkImpl[S] {
     }
 
     if (readFromWinRemain > 0) {
-      val chunk = math.min(readFromWinRemain, outRemain)
+      val chunk = math.min(readFromWinRemain, outRemain).toInt
       if (chunk > 0) {
         // logStream(s"readFromWindow(); readFromWinOff = $readFromWinOff, outOff = $outOff, chunk = $chunk")
         copyWindowToOutput(readFromWinOff = readFromWinOff, outOff = outOff, chunk = chunk)
